Now it consumes much less memory and doesn't require complicated checks to release unused memory.
Before case with "yes adkfljdsalkjflkdsalkfdsalkfldsafjdsaf" took less than 20 seconds to eat all memory, no it take 30 minutes. So, it can easily survive pikes in load.
It doesn't solve leaks for slow reader totally, only dropping info from streams can solve that.