@unclejack 2015-09-09
@LK4D4 Can you provide benchmark results, please?