@LK4D4 2015-09-10
@unclejack 
```
benchmark                       old ns/op     new ns/op     delta
Benchmark1M10BytesReads-4       319099149     123320094     -61.35%
Benchmark1M1024BytesReads-4     582125750     452420506     -22.28%
Benchmark10k32KBytesReads-4     7683824       6386629       -16.88%

benchmark                       old MB/s     new MB/s     speedup
Benchmark1M10BytesReads-4       31.34        81.09        2.59x
Benchmark1M1024BytesReads-4     1759.07      2263.38      1.29x
Benchmark10k32KBytesReads-4     42645.43     51307.19     1.20x

benchmark                       old allocs     new allocs     delta
Benchmark1M10BytesReads-4       18             16             -11.11%
Benchmark1M1024BytesReads-4     24             27             +12.50%
Benchmark10k32KBytesReads-4     21             18             -14.29%

benchmark                       old bytes     new bytes     delta
Benchmark1M10BytesReads-4       291787        317767        +8.90%
Benchmark1M1024BytesReads-4     35658444      13379362      -62.48%
Benchmark10k32KBytesReads-4     4238938       3727741       -12.06%
```
However I changed code slightly and now leak reducing isn't so good. It's little slower, but it's here. However it survives short peaks.