@unclejack 2015-09-11
@LK4D4 Those improvements are good enough. The code is also simpler now.

LGTM