@tonistiigi 2015-09-17
This PR hash been updated to do linear memory allocations. This means writes can internally go to many slices and a single read may internally read from many as well.

Note that the allocations are much higher now because we never allocate more than 1MB at a time. This could be easily changed but it helps to avoid large unused buffers in real use cases.

Bench compared to original PR:

```
benchmark                     old ns/op      new ns/op      delta
BenchmarkBytesPipeWrite       355055         362205         +2.01%
BenchmarkBytesPipeRead        101102         94173          -6.85%
Benchmark1M10BytesReads       142754210      144378924      +1.14%
Benchmark1M1024BytesReads     3425384197     1154784626     -66.29%
Benchmark10k32KBytesReads     23094954       13101353       -43.27%

benchmark                     old MB/s     new MB/s     speedup
Benchmark1M10BytesReads       70.05        69.26        0.99x
Benchmark1M1024BytesReads     298.94       886.75       2.97x
Benchmark10k32KBytesReads     14188.38     25011.16     1.76x

benchmark                     old allocs     new allocs     delta
BenchmarkBytesPipeWrite       1011           1016           +0.49%
BenchmarkBytesPipeRead        0              0              +0.00%
Benchmark1M10BytesReads       24             35             +45.83%
Benchmark1M1024BytesReads     49             1063           +2069.39%
Benchmark10k32KBytesReads     21             35             +66.67%

benchmark                     old bytes      new bytes      delta
BenchmarkBytesPipeWrite       160144         114288         -28.63%
BenchmarkBytesPipeRead        0              0              +0.00%
Benchmark1M10BytesReads       30085875       10020132       -66.69%
Benchmark1M1024BytesReads     2941720944     1029887808     -64.99%
Benchmark10k32KBytesReads     35657645       11127649       -68.79%
```

PTAL @unclejack @icecrime @aaronlehmann