@aaronlehmann 2015-09-17
:+1:. I had the same idea but didn't want to bikeshed so I didn't mention it.

Will review the code.