@aaronlehmann 2015-09-17
Some general design comments:

- I don't really like the way that `BytesPipe`'s `Read` function is nonblocking, and then gets wrapped with a `bufReader` to provide a blocking read function. I think the code would be simpler if `BytesPipe` implemented the locking and blocking logic directly.

- I think it's important to have a limit on the amount of data we can buffer, otherwise this doesn't really fix the out-of-memory situation. If there was a hard limit of, say, 1 MB per `BytesPipe`, would that cause any negative effects? I'm having trouble imaginging a situation where we need to accept huge amount of data in a nonblocking way, rather than blocking above a reasonable threshold.