@unclejack 2015-09-21
@tonistiigi Thanks for the ping. I'll be reviewing again after the latest changes.

@aaronlehmann That limits throughput greatly and it's something I've tried a while ago. We should make incremental improvements to this code because it gives us time to see how well it works for a lot of users. I'd rather not cap the buffer now. We might as well remove bufReader if we use a cap like that.