@unclejack 2015-09-22
Here are the results of the comparison between the original code and the latest version of this PR:
```
benchmark                     old ns/op      new ns/op     delta
Benchmark1M10BytesReads       212156302      139747988     -34.13%
Benchmark1M1024BytesReads     2365383852     773977845     -67.28%
Benchmark10k32KBytesReads     21738324       12166015      -44.03%

benchmark                     old MB/s     new MB/s     speedup
Benchmark1M10BytesReads       47.14        71.56        1.52x
Benchmark1M1024BytesReads     432.91       1323.04      3.06x
Benchmark10k32KBytesReads     15073.84     26934.04     1.79x

benchmark                     old allocs     new allocs     delta
Benchmark1M10BytesReads       25             35             +40.00%
Benchmark1M1024BytesReads     36             1059           +2841.67%
Benchmark10k32KBytesReads     22             35             +59.09%

benchmark                     old bytes      new bytes      delta
Benchmark1M10BytesReads       25241232       10120877       -59.90%
Benchmark1M1024BytesReads     2147493872     1031902948     -51.95%
Benchmark10k32KBytesReads     33556093       11127637       -66.84%
```

@LK4D4 @tonistiigi Do you want to address the comments regarding the documentation?