@tonistiigi 2015-09-23
@unclejack you mean the comment, or did I miss something?