@unclejack 2015-09-23
@tonistiigi Yes, that's the change I was talking about.

The code is fine, but maybe we should document BytesPipe like the drain function was documented before.