@jfrazelle 2015-09-23
@unclejack good to be merged?