@unclejack 2015-09-24
@tonistiigi Thank you for adding the documentation. We're gaining better performance and we're also keeping documentation at the same level as before on that code and this is something I appreciate.

LGTM

@jfrazelle Yes, we can move forward with this. Please feel free to review and merge whenever you can.